Split out `match_like_matches_macro`
